Create README.md
This project was created during my second year of studying Data Technology. The goal was to develop a game that runs on the lab kit, which resembles a mini gaming device. I chose to implement a Snake game where the snake grows as it eats apples. The game continues until the snake collides with its own tail, at which point it restarts